Move filters to table files
